requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
pillow==11.0.0
pip==24.3.1
sqlparse==0.5.1